First implementation to restrict target docking ports based on current vessel (fixes #4)
When switching between target docking ports and the new "Restrict Docking Ports" option is checked, only docking ports which are compatible with the current vessel are selected.  If the currently-controlled part is a docking port, then this is the only one which is checked.

Supports gendered and multi-node docking ports.