chore(build): fix Deploy project for Linux and add helper scripts
There's a bug/feature in msbuild for Linux which "helpfully" replaces
backslashes with forward slashes while parsing project files. While in
general this is useful for cross-platform compatibility with filenames,
it is a major issue when using regular expressions or, indeed, any
other situation in which a backslash is explicitly required.

See : https://github.com/dotnet/msbuild/issues/3468