feat(ui): save the window position
Thanks to the code in DMagic's "BasicOrbit" mod, added saving and
restoring the window position.

Based on research it was non-obvious how to achieve this, converting from
RectTransform to something which could be stored and back again. DMagic's
implementation is small and neat and mostly understandable and, most
importantly, it works.

The only oddity was that DMagic's code inverted the Y value when
restoring the position, which did not work. Possibly a difference in the
way the anchors are set up on the Panel object?