feat(ui): signal setting changes back to the main class
When some settings change, additional logic needs to be executed. When
the settings window was integrated in the main class it called this
logic directly whenever a button was pressed.

When the settings were decoupled, this extra logic was placed into the
TODO bin.

This commit adds an event which the main class can subscribe to whenever
a setting is changed, so that it can implement this additional logic.

It's more roundabout, but clearly separates the UI from the main logic and
will make it easier to make the settings window a native Unity UI as well.