feat(ui): refactor main class and fix a minor toolbar issue
When the toolbar button was switched from Blizzy back to Stock while the
DPAI window was open, it required 2 clicks to close the window.

Now it will toggle the window visibility regardless of what state it was
in.

Also move some of the code around and wrap it in more regions for easier
maintainability.
TODO: move more code out of the main class into helper
classes.