feat(ui): reimplement scaling of the DPAI window
Turns out it's pretty trivial using the new Unity UI - simply set the
scale of the main canvas RectTransform and everything else happens
automagically.

Also add a percentage indicator to the scaling so that a user can tell
when the scale is roughly what it's supposed to be. Perhaps we should
add "snapping" so the scale only changes in fixed increments.